8376833: Refactoring.
